LPS-63059 Move PostgreSQL provider to its own module
