+++ modules/util/portal-tools-data-partitioning-sql-builder-postgresql/bnd.bnd
@@ -1,0 +1,4 @@
+Bundle-Name: Liferay Portal Tools Data Partitioning SQL Builder PostgreSQL
+Bundle-SymbolicName: com.liferay.portal.tools.data.partitioning.sql.builder.postgresql
+Bundle-Version: 1.0.0
+Export-Package: com.liferay.portal.tools.data.partitioning.sql.builder.db.postgresql
