+++ modules/util/portal-tools-data-partitioning-sql-builder-postgresql/build.gradle
@@ -1,0 +1,3 @@
+dependencies {
+	provided project(":util:portal-tools-data-partitioning-sql-builder")
+}
