+++ modules/util/portal-tools-data-partitioning-sql-builder-postgresql/build.xml
@@ -1,0 +1,6 @@
+<?xml version="1.0"?>
+<!DOCTYPE project>
+
+<project>
+	<import file="../../build-module.xml" />
+</project>
