+++ modules/util/portal-tools-data-partitioning-sql-builder-postgresql/src/main/java/com/liferay/portal/tools/data/partitioning/sql/builder/db/postgresql/PostgreSQLProvider.java
