+++ modules/util/portal-tools-data-partitioning-sql-builder-postgresql/src/main/resources/sample-postgresql.properties
