+++ modules/util/portal-tools-data-partitioning-sql-builder-postgresql/src/test/java/com/liferay/portal/tools/data/partitioning/sql/builder/exporter/ShardExporterFactoryTest.java
@@ -14,10 +14,15 @@
 
 package com.liferay.portal.tools.data.partitioning.sql.builder.exporter;
 
-import com.liferay.portal.tools.data.partitioning.sql.builder.exporter.exception.DBProviderNotAvailableException;
+import com.liferay.portal.tools.data.partitioning.sql.builder.db.postgresql.PostgreSQLProvider;
+import com.liferay.portal.tools.data.partitioning.sql.builder.internal.util.PropsReader;
+
+import java.net.URL;
 
 import java.util.Properties;
 
+import org.junit.Assert;
+import org.junit.Assume;
 import org.junit.Test;
 
 /**
@@ -25,9 +30,24 @@ import org.junit.Test;
  */
 public class ShardExporterFactoryTest {
 
-	@Test(expected = DBProviderNotAvailableException.class)
-	public void testGetShardExporter() throws Exception {
-		ShardExporterFactory.getShardExporter(new Properties());
+	@Test
+	public void testGetShardExporterReturnsPostgreSQLProvider()
+		throws Exception {
+
+		Class<?> clazz = getClass();
+
+		URL url = clazz.getResource("/postgresql.properties");
+
+		Assume.assumeNotNull(url);
+
+		Properties properties = PropsReader.read(url.getPath());
+
+		ShardExporter shardExporter = ShardExporterFactory.getShardExporter(
+			properties);
+
+		Class<PostgreSQLProvider> providerClass = PostgreSQLProvider.class;
+
+		Assert.assertTrue(providerClass.isInstance(shardExporter));
 	}
 
 }
