+++ modules/util/portal-tools-data-partitioning-sql-builder/src/main/java/com/liferay/portal/tools/data/partitioning/sql/builder/exporter/ShardExporterFactory.java
@@ -14,7 +14,6 @@
 
 package com.liferay.portal.tools.data.partitioning.sql.builder.exporter;
 
-import com.liferay.portal.tools.data.partitioning.sql.builder.db.postgresql.PostgreSQLProvider;
 import com.liferay.portal.tools.data.partitioning.sql.builder.exporter.exception.DBProviderNotAvailableException;
 
 import java.util.Properties;
@@ -25,17 +24,6 @@ import java.util.Properties;
 public class ShardExporterFactory {
 
 	public static ShardExporter getShardExporter(Properties properties) {
-		String dataSourceClassName = properties.getProperty(
-			"dataSourceClassName", "");
-
-		if (dataSourceClassName.equals(
-				"com.impossibl.postgres.jdbc.PGDataSource") ||
-			dataSourceClassName.equals(
-				"org.postgresql.ds.PGSimpleDataSource")) {
-
-			return new PostgreSQLProvider(properties);
-		}
-
 		throw new DBProviderNotAvailableException();
 	}
